on ne vire pas adminse3 de l'annuaire......

+++ trunk/se3master/usr/share/se3/sbin/peopleClean.pl
@@ -17,7 +17,7 @@ $res = $lcs_ldap->search(base   => "$peopleDn",
 warn $res->error if $res->code;
 foreach $entry ($res->entries) {
   $uid = $entry->get_value('uid');
-  next if ($uid eq 'admin' or $uid eq 'webmaster.etab' or $uid eq 'wetab' or $uid eq 'etabw' or $uid eq 'spip.manager' or $uid eq 'unattend');
+  next if ($uid eq 'admin' or $uid eq 'adminse3' or $uid eq 'webmaster.etab' or $uid eq 'wetab' or $uid eq 'etabw' or $uid eq 'spip.manager' or $uid eq 'unattend');
   $res = $lcs_ldap->delete("uid=$uid,$peopleDn");
   print "Suppression de l'utilisateur $uid\n";
   warn $res->error if $res->code;
